- Replaced deprecated pkg_resources (backwards compatible up to 3.7) - Added Macro in setup.py to force new numpy api - Replaced code belonging to old numpy api - Fixed Build Problems on Cython>=3.0.0 and Numpy>=2.0.0 - Replaced np.npy_clongdouble with long double complex to fix compilation issues - Fixed dtypes that broke due to Numpy 2.0.0

+++ tests/run_tests.py
@@ -161,23 +161,19 @@ class TestEigency(unittest.TestCase):
 
     def test_function_type_int(self):
         # C++ long - Note that this is the standard Python integer
-        mat_in = np.array([[1, 2, 3, 4], [5, 6, 7, 8]], order="F")
-        # # equivalent to:
-        # mat_in = np.array([[1, 2, 3, 4], [5, 6, 7, 8]], order='F', dtype=np.int_)
+        mat_in = np.array([[1, 2, 3, 4], [5, 6, 7, 8]], order="F", dtype=np.int32)
         mat_out = eigency_tests.function_type_long(mat_in)
         assert_array_equal(mat_in, mat_out)
 
     def test_function_type_long(self):
         # C++ long - Note that this is the standard Python integer
-        mat_in = np.array([[1, 2, 3, 4], [5, 6, 7, 8]], order="F", dtype=int)
-        # # equivalent to:
-        # mat_in = np.array([[1, 2, 3, 4], [5, 6, 7, 8]], order='F', dtype=np.int_)
+        mat_in = np.array([[1, 2, 3, 4], [5, 6, 7, 8]], order="F", dtype=np.long)
         mat_out = eigency_tests.function_type_long(mat_in)
         assert_array_equal(mat_in, mat_out)
 
     def test_function_type_ulong(self):
         # C++ ulong
-        mat_in = np.array([[1, 2, 3, 4], [5, 6, 7, 8]], order="F", dtype=np.uint)
+        mat_in = np.array([[1, 2, 3, 4], [5, 6, 7, 8]], order="F", dtype=np.ulong)
         mat_out = eigency_tests.function_type_ulong(mat_in)
         assert_array_equal(mat_in, mat_out)
